Angularv1

+++ angular-src/src/app/components/admin-zalen/admin-zalen.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {Zaal} from "../../models/zaal.model";
 import {ZaalService} from "../../services/zaal.service";
+import { Router} from "@angular/router";
 
 @Component({
   selector: 'app-admin-zalen',
@@ -8,20 +9,27 @@ import {ZaalService} from "../../services/zaal.service";
   styleUrls: ['./admin-zalen.component.css']
 })
 export class AdminZalenComponent implements OnInit {
-  zalen:Zaal[];
+  zalen: Zaal[];
 
-  constructor(private zaalService:ZaalService) { }
-
-  ngOnInit() {
-    this.zalen = [
-      new Zaal(1, "CompUzaal 1", "CompUzaal 1 heeft 20 vaste computers en 5 laptops", 40, "http://via.placeholder.com/350x150", "0900", "2100", 20),
-      new Zaal(2, "DansZaal", "balletbar en spiegels", 40, "http://via.placeholder.com/350x150", "0900", "2100", 20),
-      new Zaal(3, "FeestZaal", "tafels en stoelen aanwezig", 40, "http://via.placeholder.com/350x150", "0900", "2100", 20),
-      new Zaal(4, "SuperZaal", "opdeelbaar in 3", 40, "http://via.placeholder.com/350x150.png", "0900", "2100", 20)
-    ];
+  constructor( private zaalService: ZaalService,private router: Router) {
   }
 
-  onDelete(id){
-    this.zaalService.deleteZaal(id);
+  ngOnInit() {
+    this.zaalService.getZalen().then(zalen=>this.zalen = zalen).then(function () {
+      this.zalen = this.zalen.sort((a, b) => {
+        if (a.naam < b.naam) return -1;
+        else if (a.naam > b.naam) return 1;
+        else return 0;
+      })
+    }.bind(this));
+  }
+
+  async onDelete(id){
+    if(confirm('Bent u zeker dat u deze zaal wil verwijderen?')){
+      await this.zaalService.deleteZaal(id);
+      this.router.navigateByUrl('/').then(()=>{
+        this.router.navigateByUrl('/adminzalen');
+      });
+    }
   }
 }
